backend/requirements.txt
# FastAPI Framework
fastapi
uvicorn[standard]

# HTTP Client for external APIs
httpx
aiofiles

# Data Validation
pydantic
pydantic-settings

# Authentication (for Supabase JWT validation)
python-jose[cryptography]
python-multipart

# Environment
python-dotenv

# Supabase client (for Stage 1-3)
supabase

# Document processing libraries
python-docx
PyPDF2
openpyxl
python-pptx

# AI/ML libraries
google-generativeai

# Development & Testing
pytest
pytest-asyncio
black
isort